+++ src/main/java/com/amazon/ion/impl/macro/MacroRef.kt
@@ -8,11 +8,12 @@ package com.amazon.ion.impl.macro
 sealed interface MacroRef {
     // TODO: See if these could be inline value classes
     @JvmInline value class ByName(val name: String) : MacroRef
-    @JvmInline value class ById(val id: Long) : MacroRef
+    // Ion is not limited to Int.MAX_VALUE macro addresses, but Java collection sizes are limited by Int.MAX_VALUE
+    @JvmInline value class ById(val id: Int) : MacroRef
     // TODO: Since system macros have an independent address space, do we need to have a `SystemById` variant?
 
     companion object {
         @JvmStatic
-        fun byId(id: Long): MacroRef = ById(id)
+        fun byId(id: Int): MacroRef = ById(id)
     }
 }
